restored edit credit modal
