+++ TODO.md
@@ -2,23 +2,13 @@
 
 ## PRIORITIES
 
-### Profile
-
-- Media: photos
-- Media: videos
-- Resume
-- ADD: "Willing to Tour" checkbox
-- Socials
-  - Website below phone number as text link
-  - Remaining social icons after Website
-  - Resume button embiggened, below socials
-
 ## TODO
 
 - Set up login cookie expirations, and a "remember me" option for a longer expiration
 - WPGraphQL Persisted Queries: <https://github.com/Quartz/wp-graphql-persisted-queries>
 - Saved searches
 
-## Trivial
+## 1.0.7
 
-- Favicon (index.html)
+- EditCreditModal: Skills not loading if secondary department and job(s) selected while the first job selected has no associated skills
+- EditCreditModal: On first load, opening a Credit shows Jobs and Skills, but Department is unselected. Subsequent openings, it works.
