+++ changelog.md
@@ -12,8 +12,7 @@ All notable changes to this project will be documented in this file, starting wi
 - Improved: Changed website URL display to "View Website" to handle for long, ugly links.
 - Improved: Minor styling tweaks
 - Improved: Credit Save/Cancel buttons no longer rerendering unncessarily
-- Improved: Credits saved without `position` terms show a reminder on Edit screen
-- Refactored: EditCreditModal component, and merged contents of EditCreditView into it.
+- Improved: Credits saved without `position` terms show a reminder on Edit scree
 
 ## \[1.0.6] - 2023-06-12 - 2023-06-13
